fix:【漏洞】修复删除下载任务时，路径遍历漏洞
# Conflicts:
#	core/core-backend/src/main/java/io/dataease/exportCenter/manage/ExportCenterManage.java

+++ core/core-backend/src/main/java/io/dataease/exportCenter/manage/ExportCenterManage.java
@@ -22,6 +22,7 @@ import io.dataease.permission.util.V3UserUtil;
 import io.dataease.result.PageResult;
 import io.dataease.system.manage.SysParameterManage;
 import io.dataease.utils.*;
+import io.dataease.visualization.dao.auto.entity.CoreStore;
 import io.dataease.visualization.dao.auto.entity.VisualizationWatermark;
 import io.dataease.visualization.dao.auto.mapper.VisualizationWatermarkRepository;
 import io.dataease.visualization.dto.WatermarkContentDTO;
@@ -86,16 +87,18 @@ public class ExportCenterManage implements BaseExportApi {
     }
 
     public void delete(String id) {
-        Iterator<Map.Entry<String, Future>> iterator = Running_Task.entrySet().iterator();
-        while (iterator.hasNext()) {
-            Map.Entry<String, Future> entry = iterator.next();
-            if (entry.getKey().equalsIgnoreCase(id)) {
-                entry.getValue().cancel(true);
-                iterator.remove();
+        if (coreExportTaskRepository.existsById(id)) {
+            Iterator<Map.Entry<String, Future>> iterator = Running_Task.entrySet().iterator();
+            while (iterator.hasNext()) {
+                Map.Entry<String, Future> entry = iterator.next();
+                if (entry.getKey().equalsIgnoreCase(id)) {
+                    entry.getValue().cancel(true);
+                    iterator.remove();
+                }
             }
+            FileUtils.deleteDirectoryRecursively(exportData_path + id);
+            coreExportTaskRepository.deleteById(id);
         }
-        FileUtils.deleteDirectoryRecursively(exportData_path + id);
-        coreExportTaskRepository.deleteById(id);
     }
 
     public void deleteAll(String type) {
